+++ pyabc/populationstrategy.py
@@ -51,7 +51,7 @@ class PopulationStrategy:
         self.nr_samples_per_parameter = nr_samples_per_parameter
 
     def adapt_population_size(self, transitions: List[Transition],
-                              model_weights: np.ndarray):
+                              model_weights: np.ndarray, t: int = None):
         """
         Select the population size for the next population.
 
@@ -59,6 +59,7 @@ class PopulationStrategy:
         ----------
         transitions: List of Transitions
         model_weights: array of model weights
+        t: Time to adapt for
 
         Returns
         -------
@@ -108,7 +109,8 @@ class ConstantPopulationSize(PopulationStrategy):
         Number of samples to draw for a proposed parameter
     """
 
-    def adapt_population_size(self, transitions, model_weights):
+    def adapt_population_size(self, transitions: List[Transition],
+                              model_weights: np.ndarray, t: int = None):
         pass
 
 
@@ -177,7 +179,7 @@ class AdaptivePopulationSize(PopulationStrategy):
                 "mean_cv": self.mean_cv}
 
     def adapt_population_size(self, transitions: List[Transition],
-                              model_weights: np.ndarray):
+                              model_weights: np.ndarray, t: int = None):
         test_X = [trans.X for trans in transitions]
         test_w = [trans.w for trans in transitions]
 
@@ -196,3 +198,30 @@ class AdaptivePopulationSize(PopulationStrategy):
 
         logger.info("Change nr particles {} -> {}"
                     .format(reference_nr_part, self.nr_particles))
+
+
+class ListPopulationSize(PopulationStrategy):
+    """
+    Return population size values from a predefined list. For every time point
+    enquired later (specified by time t), an entry must exist in the list.
+
+    Parameters
+    ----------
+    values: List[float]
+        List of population size values.
+        ``values[t]`` is the value for population t.
+    """
+
+    def __init__(self,
+                 values: List[float]):
+        super().__init__(nr_particles=list(values)[0])
+        self.population_values = list(values)
+
+    def get_config(self):
+        config = super().get_config()
+        config["population_values"] = self.population_values
+        return config
+
+    def adapt_population_size(self, transitions: List[Transition],
+                              model_weights: np.ndarray, t: int = None):
+        self.nr_particles = self.population_values[t]